Refactoring custom attributes
